refactor: migrate to local SQLite database and rewrite vulns_ui
Database & Data Layer:
- Replace TaskStore with TaskStore2 (using local uavred.db)
- Add comprehensive data models (Task, Asset, Vulnerability, etc.)
- Add sqlez and sqlez_macros crates for SQLite operations
- Add db crate for database management

UI Updates:
- Refactor kanban_ui to use new TaskStore API
- Refactor dashboard_ui for database integration
- Fix assets_ui type compatibility (AssetStatus, AssetNode)
- Rewrite vulns_ui with 3-column layout (list/detail/CVE)
- Add VulnData and related models to data crate

Configuration:
- Update Cargo.toml files with new dependencies
- Update main.rs to initialize and load TaskStore

Backup:
- Keep old TaskStore as task_store_old.rs for reference

+++ crates/assets_ui/src/asset_detail_panel/cards/risk_card.rs
@@ -10,7 +10,7 @@ pub struct RiskCard;
 
 impl RiskCard {
     pub fn render(node: &AssetNode) -> impl IntoElement {
-        let risk_color = risk_color(node.risk_score);
+        let risk_color = risk_color(node.risk_score as u8);
 
         v_flex()
             .p_3()
